Refactored Network.DNS.Encode structure
Most functions are now internal, but some are still
exported for tests via the new 'Network.DNS.Internal'
module.
